feat: added view typing
