+++ src/api/rank.api.ts
@@ -1,13 +1,5 @@
-import type { UUID } from "../models/models";
+import type { LeaderboardEntry } from "../models/models";
 import { API_URL, authHeaders, type GetToken } from "./api";
-
-export interface LeaderboardEntry {
-  rank: number;
-  userId: UUID;
-  username: string;
-  avatarUrl: string | null;
-  balance: number;
-}
 
 export async function fetchLeaderboard(
   limit: number,
