+++ src/views/LeaderboardView.tsx
@@ -6,7 +6,7 @@ import Skeleton from "@mui/material/Skeleton";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
 import { useLeaderboardQuery } from "../hooks/queries/useLeaderboardQuery";
-import type { LeaderboardEntry } from "../api/api";
+import type { LeaderboardEntry } from "../models/models";
 
 const LeaderboardLoadingSkeleton = () =>
   [0, 1, 2, 3, 4].map((i) => (
