+++ src/views/UserView.tsx
@@ -6,9 +6,9 @@ import Box from "@mui/material/Box";
 import Skeleton from "@mui/material/Skeleton";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import { fetchUserPortfolio } from "../api/api";
 import type { UUID } from "../models/models";
 import { PortfolioStakeSections } from "../components/portfolio/PortfolioStakeSections";
+import { fetchUserPortfolio } from "../api/portfolio.api";
 
 export default function UserView() {
   const { userId } = useParams<{ userId: string }>();
